Compare correct hosting IP to live nameservers
Previously the result should always have been true as the live A record and the live NS records *should* match up.

Now we properly get the *local* IP that Plesk is hosting the domain on and compare that to the live NS records.

Note: if the DNS is external this will fail every time, but that's generally OK as we don't want or need to zone-transfer those records to the secondary DNS anyway.